// src/app/models/fse-prospect-details.model.ts
export interface Prospect {
    saveFlag: boolean;
    updateFlag: boolean;
    deleteFlag: boolean;
    getFlag: boolean;
    fse_angaza_id: string;
    prospect_id: string;
    status: string;
    status_updated_at: string;
    ticket_id: string;
    approved: string;
    message: string;
    attempt: string;
    country: string;
  }  